requirements.txt
yfinance>=0.2.36
pandas>=2.0
openpyxl>=3.1
xlrd>=2.0
rich>=13.0
python-dotenv>=1.0
finnhub-python>=2.4
